Face_Tagger/requirements.txt
pysimplegui
PIL
Panda
Openpyxl 